Se creo la interfaz renovable y se implemento en la clase libro, ademas se movio la implementacion de prestable a la clase RecursosDigitales
