+++ interfaces/Renovable.java
@@ -1,0 +1,5 @@
+package interfaces;
+
+public interface Renovable {
+    void renovar();
+}
