+++ modelos/RecursoDigital.java
@@ -1,12 +1,14 @@
 package modelos;
 
 import interfaces.InterfazRD;
+import interfaces.Prestable;
 
 import java.util.Scanner;
 
-public abstract class RecursoDigital implements InterfazRD {
+public abstract class RecursoDigital implements InterfazRD, Prestable {
     protected String titulo;
     protected String autor;
+    protected boolean prestado;
 
     public RecursoDigital(String titulo, String autor) {
         setTitulo(titulo);
@@ -41,5 +43,25 @@ public abstract class RecursoDigital implements InterfazRD {
     public String toString() {
         return "Titulo: " + titulo +
                 "\nAutor: " + autor;
+    }
+
+    @Override
+    public void prestar() {
+        if (!prestado) {
+            prestado = true;
+            System.out.println("El recurso fue prestado correctamente");
+        }else {
+            System.out.println("El recurso ya esta prestado");
+        }
+    }
+
+    @Override
+    public void devolver() {
+        if (prestado) {
+            prestado = false;
+            System.out.println("El recurso fue devuelto correctamente");
+        } else {
+            System.out.println("El recurso ya esta devuelto");
+        }
     }
 }
